requirements.txt
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
appdirs==1.4.3
Django==1.11.1
django-crispy-forms==1.6.1
django-filter==1.0.4
django-guardian==1.4.8
djangorestframework==3.6.3
Markdown==2.6.8
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
